rose/src/com/internousdev/rose/dto/PurchaseHistoryInfoDTO.java
package com.internousdev.rose.dto;

import java.util.Date;

public class PurchaseHistoryInfoDTO {
	private int id;
//	購入履歴情報
	private String userId;
	private int productId;
	private int productCount;
	private int price;
	private int destinationId;
	private Date registDate;
	private Date updateDate;
//	商品情報
	private String productName;
	private String productNameKana;
	private String imageFilePath;
	private String imageFileName;
	private String releaseCompany;
	private Date releaseDate;
//その他
	private int totalPrice;//	合計金額（商品価格　＊　個数）
	private int subTotal;

//	ゲッター、セッター
	public int getId() {
		return id;
	}
	public void setId(int id) {
		this.id = id;
	}

	public String getUserId() {
		return userId;
	}
	public void setUserId(String userId) {
		this.userId = userId;
	}

	public int getProductId() {
		return productId;
	}
	public void setProductId(int productId) {
		this.productId = productId;
	}

	public int getProductCount() {
		return productCount;
	}
	public void setProductCount(int productCount) {
		this.productCount = productCount;
	}

	public int getPrice() {
		return price;
	}
	public void setPrice(int price) {
		this.price=price;
	}

	public int getDestinationId() {
		return destinationId;
	}
	public void setDestinationId(int destinationId) {
		this.destinationId = destinationId;
	}

	public Date getRegistDate() {
		return registDate;
	}
	public void setRegistDate(Date registDate) {
		this.registDate = registDate;
	}

	public Date getUpdateDate() {
		return updateDate;
	}
	public void setUpdateDate(Date updateDate) {
		this.updateDate = updateDate;
	}

	public String getProductName() {
		return productName;
	}
	public void setProductName(String productName) {
		this.productName = productName;
	}

	public String getProductNameKana() {
		return productNameKana;
	}
	public void setProductNameKana(String productNameKana) {
		this.productNameKana = productNameKana;
	}

	public String getImageFilePath() {
		return imageFilePath;
	}
	public void setImageFilePath(String imageFilePath) {
		this.imageFilePath = imageFilePath;
	}

	public String getImageFileName() {
		return imageFileName;
	}
	public void setImageFileName(String imageFileName) {
		this.imageFileName = imageFileName;
	}

public String getReleaseCompany() {
	return releaseCompany;
}
public void setReleaseCompany(String releaseCompany) {
	this.releaseCompany = releaseCompany;
}

public Date getReleaseDate() {
	return releaseDate;
}
public void setReleaseDate(Date releaseDate) {
	this.releaseDate = releaseDate;
}
	public int getTotalPrice() {
		return totalPrice;
	}
	public void setTotalPrice(int totalPrice) {
		this.totalPrice = totalPrice;
	}

	public int getSubTotal() {
		return subTotal;
	}
	public void setSubTotal(int subTotal) {
		this.subTotal = subTotal;
	}

}
